Misc fixes to misc fixes
