+++ addons/miscFixes/CfgMagazines.hpp
@@ -1,0 +1,9 @@
+class CfgMagazines {
+    class CA_LauncherMagazine;
+    class rhs_mag_smaw_HEAA: CA_LauncherMagazine {
+        mass = 80;
+    };
+    class rhs_mag_smaw_HEDP: CA_LauncherMagazine {
+        mass = 65;
+    };
+};
